Refactor UI from data access in acars
